divide record iterator class
