+++ include/bztree/bztree.hpp
@@ -31,6 +31,7 @@
 
 // local sources
 #include "bztree/component/node.hpp"
+#include "bztree/component/record_iterator.hpp"
 #include "bztree/utility.hpp"
 
 namespace dbgroup::index::bztree
@@ -49,6 +50,7 @@ class BzTree
   using Metadata = component::Metadata;
   using StatusWord = component::StatusWord;
   using Node_t = component::Node<Key, Compare>;
+  using RecordIterator_t = component::RecordIterator<Key, Payload, Compare>;
   using NodeRC = component::NodeRC;
   using NodeGC_t = ::dbgroup::memory::EpochBasedGC<Page>;
   using NodeStack = std::vector<std::pair<Node_t *, size_t>>;
@@ -64,173 +66,6 @@ class BzTree
   using BulkFuture = std::future<BulkResult>;
 
  public:
-  /*####################################################################################
-   * Public classes
-   *##################################################################################*/
-
-  /**
-   * @brief A class to represent a iterator for scan results.
-   *
-   * @tparam Key a target key class
-   * @tparam Payload a target payload class
-   * @tparam Compare a key-comparator class
-   */
-  class RecordIterator
-  {
-    using BzTree_t = BzTree<Key, Payload, Compare>;
-
-   public:
-    /*##################################################################################
-     * Public constructors and assignment operators
-     *################################################################################*/
-
-    RecordIterator(  //
-        BzTree_t *bztree,
-        Node_t *node,
-        const size_t begin_pos,
-        const size_t end_pos,
-        const ScanKey end_key,
-        const bool is_right_end)
-        : bztree_{bztree},
-          node_{node},
-          record_count_{end_pos},
-          current_pos_{begin_pos},
-          current_meta_{node->GetMetadata(current_pos_)},
-          end_key_{std::move(end_key)},
-          is_right_end_{is_right_end}
-    {
-    }
-
-    constexpr RecordIterator &
-    operator=(RecordIterator &&obj) noexcept
-    {
-      node_ = obj.node_;
-      record_count_ = obj.record_count_;
-      current_pos_ = obj.current_pos_;
-      current_meta_ = obj.current_meta_;
-      is_right_end_ = obj.is_right_end_;
-
-      return *this;
-    }
-
-    RecordIterator(const RecordIterator &) = delete;
-    RecordIterator &operator=(const RecordIterator &) = delete;
-    RecordIterator(RecordIterator &&) = delete;
-
-    /*##################################################################################
-     * Public destructors
-     *################################################################################*/
-
-    ~RecordIterator() = default;
-
-    /*##################################################################################
-     * Public operators for iterators
-     *################################################################################*/
-
-    /**
-     * @retval true if this iterator indicates a live record.
-     * @retval false otherwise.
-     */
-    explicit
-    operator bool()
-    {
-      return HasRecord();
-    }
-
-    /**
-     * @return a current key and payload pair
-     */
-    constexpr auto
-    operator*() const  //
-        -> std::pair<Key, Payload>
-    {
-      return {GetKey(), GetPayload()};
-    }
-
-    /**
-     * @brief Forward an iterator.
-     *
-     */
-    constexpr void
-    operator++()
-    {
-      ++current_pos_;
-      current_meta_ = node_->GetMetadata(current_pos_);
-    }
-
-    /*##################################################################################
-     * Public getters/setters
-     *################################################################################*/
-
-    /**
-     * @brief Check if there are any records left.
-     *
-     * function may call a scan function internally to get a next leaf node.
-     *
-     * @retval true if there are any records or next node left.
-     * @retval false if there are no records and node left.
-     */
-    auto
-    HasRecord()  //
-        -> bool
-    {
-      while (true) {
-        if (current_pos_ < record_count_) return true;  // records remain in this node
-        if (is_right_end_) return false;                // this node is the end of range-scan
-
-        // update this iterator with the next scan results
-        const auto &next_key = node_->GetHighKey();
-        *this = bztree_->Scan(std::make_tuple(next_key, 0, kClosed), end_key_);
-      }
-    }
-
-    /**
-     * @return a key of a current record
-     */
-    [[nodiscard]] constexpr auto
-    GetKey() const  //
-        -> Key
-    {
-      return node_->GetKey(current_meta_);
-    }
-
-    /**
-     * @return a payload of a current record
-     */
-    [[nodiscard]] constexpr auto
-    GetPayload() const  //
-        -> Payload
-    {
-      return node_->template GetPayload<Payload>(current_meta_);
-    }
-
-   private:
-    /*##################################################################################
-     * Internal member variables
-     *################################################################################*/
-
-    /// a pointer to BwTree to perform continuous scan
-    BzTree_t *bztree_{nullptr};
-
-    /// the pointer to a node that includes partial scan results
-    Node_t *node_{nullptr};
-
-    /// the number of records in this node
-    size_t record_count_{0};
-
-    /// the position of a current record
-    size_t current_pos_{0};
-
-    /// the metadata of a current record
-    Metadata current_meta_{};
-
-    /// the end key given from a user
-    ScanKey end_key_{};
-
-    /// a flag for indicating whether scan has finished
-    bool is_right_end_{};
-  };
-
   /*####################################################################################
    * Public constructors and assignment operators
    *##################################################################################*/
@@ -312,7 +147,7 @@ class BzTree
   Scan(  //
       const ScanKey &begin_key = std::nullopt,
       const ScanKey &end_key = std::nullopt)  //
-      -> RecordIterator
+      -> RecordIterator_t
   {
     [[maybe_unused]] const auto &guard = gc_.CreateEpochGuard();
 
@@ -338,7 +173,7 @@ class BzTree
     // check the end position of scanning
     const auto [is_end, end_pos] = page->SearchEndPositionFor(end_key);
 
-    return RecordIterator{this, page.get(), begin_pos, end_pos, end_key, is_end};
+    return RecordIterator_t{this, page.get(), begin_pos, end_pos, end_key, is_end};
   }
 
   /*####################################################################################
